Updating (changes and flask)
Added two files that start with "other", which are essentially the same files as the others, but with minor modifications and the images for the MultinomialNB model. I uploaded screenshots of those to the static file.  Some of the variables have changed names becuase of I get issues when pulling from the main branch so I apologize in advance. I'll submit screenshots of the app to show what it looks like